Rename file to match exported member
