+++ gatsby/gatsby-ts/gatsby-node/index.ts
@@ -1,3 +1,3 @@
 export * from './createPages';
 export * from './onCreateWebpackConfig';
-export * from './onPostBuild';
+export * from './onPreBuild';
